requirements.txt
appnope==0.1.4
asttokens==3.0.0
certifi==2025.1.31
comm==0.2.2
debugpy==1.8.12
decorator==5.2.0
executing==2.2.0
filelock==3.24.3
fsspec==2026.2.0
ipykernel==6.29.5
ipython==8.32.0
jedi==0.19.2
Jinja2==3.1.6
jupyter_client==8.6.3
jupyter_core==5.7.2
MarkupSafe==3.0.3
matplotlib-inline==0.1.7
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.6.1
numpy==2.4.2
packaging==24.2
parso==0.8.4
pexpect==4.9.0
pillow==12.1.1
platformdirs==4.3.6
prompt_toolkit==3.0.50
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.1
python-dateutil==2.9.0.post0
pyzmq==26.2.1
setuptools==82.0.0
six==1.17.0
stack-data==0.6.3
sympy==1.14.0
torch==2.10.0
torchvision==0.25.0
tornado==6.4.2
traitlets==5.14.3
typing_extensions==4.15.0
wcwidth==0.2.13
